00-TALK_TO_MY_BOOK/requirements.txt
python-dotenv
chromadb==0.5.7
langchain-chroma==0.1.4
langchain-community==0.3.1
langchain-core==0.3.6
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
sentence-transformers==3.1.1
streamlit
transformers==4.45.2
typing==3.7.4.3